Make ::label_order implicit via ::describe_label
Previously label_order was explicitly declared.  This looked ugly and
allowed funny ordering of information in the console.  Now the label
ordering is implicit through the order of calls to
PowerMitten::Task::describe_label.

This means the Task labels will always appear to the left of any labels
for a subclass which is good for consistency.  It also makes it easier
to define what you want to show in the console for a task, no extra call
to manage the label order.