Replace Struct with Resource
Resource is very Struct-like but more focused for these needs.  I'm
unsure if it is the best way to go, though.

A Resource must be told a little about what the fields in the JSON
object are so it can create the object appropriately.  It may be more
appropriate to automatically detecting structures at load time, but the
API for for OpenStack doesn't allow discovery of types at this time.

A drawback to Resource compared to Struct is that you must give it a
class name, but this is minor.  This could be worked-around through an
extra dynamic constant lookup, but it seems too minor to "fix".