Move queue console stats to PowerMitten::Queue
This makes Queue uniform with the other services (except Mutex) in that
the console does not have to treat it specially.  This change causes
PowerMitten::Queue to forward calls to a wrapped Queue instance.  This
is less than desirable as it will reduce performance somewhat.

Perhaps a future change can make PowerMitten::Task a module