Move mutex console stats to PowerMitten::Mutex
Like PowerMitten::Queue, PowerMitten::Mutex uses Forwardable to wrap a
Mutex.

This also removes Control#add_service.  The #add_* methods need to be
refactored because they are all nearly identical.

This also works around a bug when forking the Mutex instance.  For some
reason the forked process has a ThreadError about a locked mutex when it
sleeps.