Remove uniform code dependency from the console
Previously the console would have to be run where all processes had the
same version of the code.  An update to one part could cause the console
to break.

This restriction has been removed by including a reference to the
remote object's class in the description instead of the class.

This changed removed the console's caching of the column descriptions as
they can no longer be tied to a single class.